Rework folder request response structure
