+++ crates/bitwarden/src/mobile/vault/folders/folder_encrypt.rs
@@ -11,12 +11,6 @@ pub struct FolderEncryptRequest {
 
 #[derive(Serialize, Deserialize, Debug, JsonSchema)]
 #[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub struct FolderDecryptRequest {
+pub struct FolderEncryptResponse {
     pub folder: Folder,
-}
-
-#[derive(Serialize, Deserialize, Debug, JsonSchema)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub struct FolderDecryptListRequest {
-    pub folders: Vec<Folder>,
 }
