+++ crates/bitwarden/src/mobile/vault/folders/mod.rs
@@ -1,5 +1,7 @@
-mod request;
-mod response;
+mod folder_decrypt;
+mod folder_decrypt_list;
+mod folder_encrypt;
 
-pub use request::{FolderDecryptListRequest, FolderDecryptRequest, FolderEncryptRequest};
-pub use response::{FolderDecryptListResponse, FolderDecryptResponse, FolderEncryptResponse};
+pub use folder_decrypt::{FolderDecryptRequest, FolderDecryptResponse};
+pub use folder_decrypt_list::{FolderDecryptListRequest, FolderDecryptListResponse};
+pub use folder_encrypt::{FolderEncryptRequest, FolderEncryptResponse};
